requirements.txt
Flask==2.1.1
requests==2.26.0
langchain
duckduckgo_search
python-dotenv
Werkzeug==2.0.3

